requirements.txt
django-filter
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
django-heroku==0.3.1